Fix repetition in flask and django clients
